W temp

+++ GameEntity/Tank.h
@@ -31,7 +31,7 @@ private:
 protected:
     Tank(
         const double& health, const double& health_regen, const double& max_health, 
-        const int& size, const int& vx, const int& vy,const double& xp,
+        const int& size, const double& vx, const double& vy,const double& xp,
         const double& attack_speed,
         const double& bullet_speed,
         const double& damage,
